Move GdgActivity into common package
