+++ app/src/main/java/org/gdg/frisbee/android/about/AboutActivity.java
@@ -21,7 +21,7 @@ import android.support.v4.view.ViewPager;
 import android.view.MenuItem;
 
 import org.gdg.frisbee.android.R;
-import org.gdg.frisbee.android.activity.GdgActivity;
+import org.gdg.frisbee.android.common.GdgActivity;
 import org.gdg.frisbee.android.widget.SlidingTabLayout;
 
 import butterknife.InjectView;
