+++ app/src/main/java/org/gdg/frisbee/android/common/GdgActivity.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.gdg.frisbee.android.activity;
+package org.gdg.frisbee.android.common;
 
 import android.app.ActivityManager;
 import android.app.PendingIntent;
@@ -33,6 +33,7 @@ import com.google.android.gms.plus.Plus;
 
 import org.gdg.frisbee.android.R;
 import org.gdg.frisbee.android.achievements.AchievementActionHandler;
+import org.gdg.frisbee.android.activity.TrackableActivity;
 import org.gdg.frisbee.android.utils.PrefUtils;
 import org.gdg.frisbee.android.utils.ScopedBus;
 import org.gdg.frisbee.android.utils.Utils;
