+++ app/src/main/java/org/gdg/frisbee/android/event/EventActivity.java
@@ -26,7 +26,7 @@ import android.view.MenuItem;
 
 import org.gdg.frisbee.android.Const;
 import org.gdg.frisbee.android.R;
-import org.gdg.frisbee.android.activity.GdgActivity;
+import org.gdg.frisbee.android.common.GdgActivity;
 import org.gdg.frisbee.android.widget.SlidingTabLayout;
 
 import butterknife.InjectView;
